requirements.txt
rdflib==7.0.0
pyshacl==0.24.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.26
pandas>=2.0
matplotlib>=3.8
transformers>=4.43.0
torch>=2.3.0
